@@ -8,8 +8,8 @@ import { ElementsDataSource } from './elements-datasource';
   styleUrls: ['./elements.component.css']
 })
 export class ElementsComponent implements OnInit {
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
   dataSource: ElementsDataSource;
 
   /** Columns displayed in the table. Columns IDs can be added, removed, or reordered. */
